\name{RcppJson-package}
\alias{RcppJson-package}
\alias{RcppJson}
\docType{package}
\title{\packageTitle{RcppJson}}
\description{\packageDescription{RcppJson}}
\details{The DESCRIPTION file: \packageDESCRIPTION{RcppJson}}
\section{Package Content}{\packageIndices{RcppJson}}
\author{\packageAuthor{RcppJson}}
\section{Maintainer}{\packageMaintainer{RcppJson}}
\references{http://www.boost.org/}
\keyword{ package }
